requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
cloudinary==1.22.0
confusable-homoglyphs==3.2.0
Django==3.1.3
django-bootstrap4==2.3.1
django-registration==3.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
PyJWT==1.7.1
pytz==2020.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.0
